+++ src/core/memory.h
@@ -1,6 +1,9 @@
 #pragma once
 
 #include <string.h>
+
+#include "common.h"
+
 #define mem_malloc(size) mem_reallocate(NULL, size)
 #define mem_realloc(ptr, size) mem_reallocate(ptr, size)
 #define mem_free(ptr) mem_reallocate(ptr, 0)
@@ -15,4 +18,4 @@
 #define mem_zero(array, T, count) memset(array, 0, sizeof(T) * (count))
 #define mem_copy(dest, orig, size) memcpy(dest, orig, size)
 
-void *mem_reallocate(void *buffer, int new_capacity);
+void *mem_reallocate(void *buffer, usize new_capacity);
